Reenabled bonds
some type-changes:
- Atom has separate change-tracker for coordinates and rest
- structs for display-data
- fitted some data to directly fit these structs

fixed formatting

+++ libvipster/bond.h
@@ -6,9 +6,9 @@ namespace Vipster {
         std::size_t at1;
         std::size_t at2;
         float dist;
-        long xdiff;
-        long ydiff;
-        long zdiff;
+        int16_t xdiff;
+        int16_t ydiff;
+        int16_t zdiff;
     };
 }
